raftstore: extract peer_storage helpers
